Add pbft_raft_safe_live notebook
1) Why is the change being made?
---------------------------------
Now that we have the theory for calculating pbft/raft safety/liveness we
can easily calculate how differing quorum sizes compare for
safety/liveness. Adding this notebook allows anyone to investigate how
differing quorums and probability of failure change deployment
characteristics. It also allows people to read about why the thresholds
for safety/liveness are what they are.

2) What has changed?
---------------------------------
A noteboook has been added to calculate the chance of the system being
safe/live for pbft and raft. This uses the epoch model where we see the
probability of enough node failure occuring to make unsafe/unlive
executions possible.

3) Other concerns and/or dependent changes?
---------------------------------
This is still calculating the probability of safety/liveness using an
epoch based approach, not a markov chain. We can extend this notebook to
use this approach to assign probabilities for different strategies as
opposed to the current binomial approach.

Additionally -- the bulk of the code in the notebook should be moved to
a seperate .py file in the future to make the notebook cleaner.